// apps/frontend/src/types/schema.ts
export interface Database {
// ...
  public: {
    Tables: {
// ...
      horserace_payments: {
        Row: {
          created_at: string | null
          id: number
          race: number
          racecourse: number
        }
        Insert: {
          created_at?: string | null
          id?: number
          race: number
          racecourse: number
        }
        Update: {
          created_at?: string | null
          id?: number
          race?: number
          racecourse?: number
        }
        Relationships: [
          {
            foreignKeyName: "horserace_payments_race_fkey"
            columns: ["race"]
            referencedRelation: "race"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "horserace_payments_racecourse_fkey"
            columns: ["racecourse"]
            referencedRelation: "racecourse"
            referencedColumns: ["id"]
          }
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}